Send winnings to sender instead of into Tornado
Reworked so winnings are sent back, instead of having a deposit method that takes
a Tornado note.

This makes Thresher more generic, removes the dependency on the Tornado API,
and should be easier for users to understand and work with the Tornado GUI.

TODO: handle gas! The user should get 0.1 ETH plus enough to cover the
gas for a Tornado deposit (1 million gas).