Pay out enough ETH to cover gas cost of a tornado deposit
Keep track of the gasprice used for deposits, and when paying out, pay out
the tornado amount PLUS enough ETH (at the winning deposit's gas price) to
make the deposit.

Extremely high gas prices are disallowed (the payout cannot be more than
two times the payoutThreshold). And the chances of winning are based on
the amount deposited into the contract compared to the payout, so
if you deposit with a high gasprice you have less chance of winning versus
somebody who deposits the same amount with a low gasprice.