Generalized to be non-Tornado-specific
This is now a generic "little rolling lottery".

Deposit any amount and have a chance to win up to some set-at-deployment-time
maximum amount.

This is still intended to empty out temporary addresses
used for private Tornado withdrawals/deposits. A Tornado-friendly wallet
would use this to sweep up 'loose change' less than 0.1 ETH, calling
deposit(0.1ETH + tornado deposit gas cost) so either the loose change
is lost OR enough ETH is won to fund a Tornado deposit.

But now this same contract could be used for "deposit for a chance to win"
games up to 4 ETH.