Reject entries if contract balance is too low to pay out
Require that the contract have a balance of at least 'winAmount'
when accepting a new entry.

It is still possible for entries to get 'stuck' because the
contract doesn't have enough funds to pay out. For example,
if the contract starts with 9ETH, then eleven people
contribute-0.1-to-win-1, there is a one-in-10^10 chance
(hundred billion) the contract will be empty for the eleventh person.
(the first ten win 0.9 ETH each, draining the contract balance)

To be utterly safe would involve a loop through pending entries
and adding up what they could potentially win (or keeping track
of that in contract state to save gas). But doing that would
open up a denial-of-service attack; I could send
contribute-1-wei-to-win-1-ETH entries until the code decided
there was a chance it wouldn’t be able to pay out.