Rework how processAll() and contribute() deal with gas
When testing on Rinkeby I ran into situations where processAll()
would fail due to running out of gas.

So I measured gas usage in the various bits of the contract code
(by adding a temporary TraceGas event) and updated the magic
constants that I use to make sure processAll() doesn't run out
of gas, and added a unit test to exercise the code.

I also decided to transfer to winners with 50,000 gas, which should
be enough if they are entering using a simple smart contract instead of
a dumb address.